refactor: move setupBridges down
